% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_jags_latents.R
\name{export_jags_latents}
\alias{export_jags_latents}
\title{Extract latent variables from a JAGs model}
\usage{
export_jags_latents(jagsmod)
}
\arguments{
\item{jagsmod}{a runjags JAGs model}
}
\value{
the factor scores for the latent variables
}
\description{
Extract latent variables from a JAGs model
}
