% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexplavaan.R
\name{flexplavaan}
\alias{flexplavaan}
\title{Fit a flexplavaan object}
\usage{
flexplavaan(model = NULL, data = NULL, ...)
}
\arguments{
\item{model}{The model specified using lavaan syntax.}

\item{data}{The dataset supplied by the user.}

\item{...}{Other arguments passed to lavaan.}
}
\value{
A flexplavaan object with the following objects: lavaan, data, and standard_errors
}
\description{
For efficiency, I recommend against fitting lavaan objects and instead fit
flexplavaan objects. These use identical syntax to lavaan, but also computes
standard errors. Computing standard errors is computationally intensive, so this
function inserts the standard errors within a flexplavaan class so it can be reused
(without being recomputed) later.
}
