% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_implied_measurement.R
\name{implied_measurement}
\alias{implied_measurement}
\title{Display the model-implied measurement plot}
\usage{
implied_measurement(
  model,
  model2 = NULL,
  latent = NULL,
  limit = 4,
  sort_slopes = T,
  plot_residuals = T,
  ...
)
}
\arguments{
\item{model}{A lavaan object}

\item{latent}{a string, specifying which latent variable is plotted on the y-axis. 
By default, it will display all the variables and return a list of plots.}

\item{limit}{The maximum number of observed variables displayed. Defaults to 4.}

\item{sort_slopes}{Should the plots by sorted by how much their model-implied slopes 
deviate from a standard regression model slopes? Defaults to true.}

\item{...}{Other parameters passed to flexplot.}
}
\value{
Either a ggplot2 plot, or a list of ggplot2 plots
}
\description{
These models plot the latent variable on the Y-axis, and the standardized observed
variables on the x axis. Each variable is shown as a separate panel. By default, 
the algorithm limits the number of plots to 4, but the user can specify more.
}
\details{
The model-implied fit is computed using the correlations between the observed/latent
variables. Since all variables are standardized before plotting, the correlations become
the slopes, and the intercepts are set to zero.
}
