% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modifiers.R
\name{modify_smooth}
\alias{modify_smooth}
\title{Modify the default smoothing function}
\usage{
modify_smooth(p, method = "lm", se = F)
}
\arguments{
\item{p}{a plot to modify}

\item{method}{method to modify from flexplot}

\item{se}{should standard errors be reported? I'm not sure if this will actually do anything though....}
}
\value{
a ggplot object
}
\description{
Modify the default smoothing function
}
