% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_sem.R
\docType{data}
\name{mugglevwizard}
\alias{mugglevwizard}
\title{Wizard/Witch Propensity Scores for 500 Students at Hogwarts}
\format{A data frame with 500 rows and 7 variables:
\describe{
  \item{strange}{Strange occurance scores on a 40 item survey, completed by an independent observer}
  \item{telekenesis}{Number of occurances reported by potential witch/wizard, self-reported}
  \item{wingardium}{Number of seconds the student was able to levitate an object using the spell "wingardium leviosa"}
  \item{darkarts}{Score at the final year of Hogwarts in "Defense Against the Dark Arts"}
  \item{potions}{Score at the final year of Hogwarts in "Potions"}
  \item{history}{Score at the final year of Hogwarts in "History of Magic"}
  \item{mugglevwizard}{Indicator variable that identifies each student as muggle or witch/wizard}
  
}}
\source{
These data are proprietary and obtained through the school of Hogwarts.
}
\usage{
mugglevwizard
}
\description{
This dataset contains scores on various measures of wizarding ability
and their subsequent OWL scores at Hogwarts
}
\keyword{datasets}
