% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{residual_plots}
\alias{residual_plots}
\title{Plot a "hopper" plot (or residual plot) for a latent variable model}
\usage{
residual_plots(fitted, fitted2 = NULL, max_val = 0.01)
}
\arguments{
\item{fitted}{A lavaan object that is fitted}

\item{fitted2}{An optional second option we can compare the first model against}

\item{max_val}{The value at which we stop showing correlations.}
}
\value{
A hopper plot.
}
\description{
These plots show lines representing the absolute value (and negative absolute value)
of the residuals from a lavaan model. The dots indicate that actual value of the residual. 
These help identify which correlations the model failed to capture.
}
\examples{
hopper_plot(force_fit)
}
