% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visualize.R
\name{visualize.flexplavaan}
\alias{visualize.flexplavaan}
\alias{visualize}
\title{Visualize a flexplavaan model}
\usage{
\method{visualize}{flexplavaan}(
  object,
  object2 = NULL,
  subset = NULL,
  plot = "all",
  formula = NULL,
  ...
)
}
\description{
Visualize a flexplavaan model
}
