% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_sem.R
\name{visualize.lavaan}
\alias{visualize.lavaan}
\title{Visualize a linear lavaan model}
\usage{
\method{visualize}{lavaan}(object, plot = c("all", "residuals", "model"),
  formula = NULL, ...)
}
\arguments{
\item{object}{a lavaan object}

\item{plot}{what should be plotted? User can specify "diagnostics" or "model"}

\item{...}{Other arguments passed to flexplot}
}
\description{
This function generates diagnostic plots or model plots of a lavaan object.
}
