% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize a fitted model}
\usage{
visualize(
  object,
  object2 = NULL,
  subset = NULL,
  plot = c("all", "disturbance", "model", "measurement", "latent"),
  formula = NULL,
  sort_plots = TRUE,
  ...
)
}
\description{
Visualize a fitted model
}
