% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visualize.R
\name{visualize.runjags}
\alias{visualize.runjags}
\alias{visualize}
\title{Visualize a runjags model}
\usage{
\method{visualize}{runjags}(
  object,
  summary_bayesian = NULL,
  data,
  which.latent = 1,
  object2 = NULL,
  subset = NULL,
  plot = c("all", "residuals", "model"),
  formula = NULL,
  ...
)
}
\arguments{
\item{object}{a runjags object}

\item{object2}{a second object (optional). This is used to visually compare
the fit of two different models.}

\item{plot}{what should be plotted? User can specify "diagnostics" or "model"}

\item{...}{Other arguments passed to flexplot}
}
\description{
This function generates diagnostic plots or model plots of a lavaan object.
}
