% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_sem.R
\name{viz_diagnostics}
\alias{viz_diagnostics}
\title{Visualize Diagnostics for a lavaan object}
\usage{
viz_diagnostics(data, mapping, fit.lavaan, alpha = 0.5,
  plot = c("trace", "disturbance"), ...)
}
\arguments{
\item{data}{The dataset containing the raw (item) observations}

\item{mapping}{A ggplot-defined mapping (e.g., \code{aes(x=x, y=y)}).}

\item{fit.lavaan}{A fitted \code{\link{lavaan::lavaan}} object}

\item{alpha}{The transparency of the datapoints in the scatterplot. Defaults to 0.5}

\item{plot}{One of the following: "trace" or "disturbance"}

\item{...}{Other arguments passed to \code{\link{ggplot2::ggplot}}}
}
\value{
a plot of the desired relationship
}
\description{
This function visualizes pairwise relationships between items in a lavaan
object. The user can specify whether to display "trace" or "disturbance" plots 
Trace plots will show the scatterplot between pairs of items (defined via the mapping argument)
with the model-implied fit overlaying the raw data. If the line fails to pass through the more
concentrated parts of the data, this will indicate misfit. The disturbance plots (or
disturbance-dependence plots) will show the relationship between the two items after 
removing the model-implied effects. If any relationship remains, it suggests the model failed
to capture important information.
}
\examples{

}
