% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.flexplavaan}
\alias{visualize.flexplavaan}
\alias{visualize}
\title{Visualize a flexplavaan model}
\usage{
\method{visualize}{flexplavaan}(
  object,
  object2 = NULL,
  subset = NULL,
  plot = c("all", "disturbance", "model", "measurement", "latent"),
  formula = NULL,
  sort_plots = TRUE,
  model_names = NULL,
  ...
)
}
\description{
Visualize a flexplavaan model
}
