% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visualize.R
\name{visualize.lavaan}
\alias{visualize.lavaan}
\alias{visualize}
\title{Visualize a linear lavaan model}
\usage{
\method{visualize}{lavaan}(
  object,
  object2 = NULL,
  subset = NULL,
  plot = c("all", "disturbance", "model", "measurement", "latent", "residuals"),
  formula = NULL,
  sort_plots = TRUE,
  model_names = NULL,
  ...
)
}
\arguments{
\item{object}{a lavaan object}

\item{object2}{a second lavaan object (optional). This is used to visually compare
the fit of two different models.}

\item{subset}{Either a vector of numbers or variable names indicating which variables
should be plotted. Defaults to NULL}

\item{plot}{what should be plotted? User can specify "all" (default), "disturbance", 
"model", "measurement", "latent", or "residual"}

\item{formula}{For latent plots, the user can specify a \link[flexplot]{flexplot} formula. 
Option is ignored for the other plots}

\item{sort_plots}{Should the axes be sorted according to the size of the residuals? Setting to 
TRUE (default) will plot the variables with the largest residuals first}

\item{model_names}{What should the legend be named for the two lines? Defaults to NULL.}

\item{...}{Other arguments passed to flexplot}
}
\description{
This function generates diagnostic plots or model plots of a lavaan object.
}
\examples{
require(lavaan)
data("correct_small")

model = "
f1 =~ x1 + x2 + x3
f2 =~ y1 + y2 + y3
f1 ~ f2
"
  fit.lavaan = cfa(model, data=correct_small)
  visualize(fit.lavaan)
}
